Cleanup and better format checks.
Changed super block to indicate the number of sequential zones to be reserved
for reclaim. This is configurable using the command line. The default is 16
zones.

The command line interface was also changed to facilitate parsing: the operation
comes first, followed by the device path and the options for the operation.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>