Fix double free
In case of failure to obtain a device zone information, freeing of the
device zone array is done twice, once in dmz_get_dev_zones() and a
second time in dmz_close_dev(), leading to a double free error.

Fix this by removing the call to free in dmz_get_dev_zones() and setting
the zones pointer to NULL after free in dmz_close_dev().

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>