dmzadm: Fix target activation
When using a kernel version before 5.8, that is, a dm-zoned kernel
module V1, activating the target (wrongly) require that the target
length be the entire backend device capacity. Use dmz_mod_ver to detect
V1 modules and fix the target capacity specified in the target entry for
this older version.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>