dmzadm: fix metadata version handling on start
When starting a dm-zoned disk formatted with metadata version 1, make
sure to not use the UUID to start the disk as this older metadata
version did not set a UUID field in the superblock. This fixes a start
failure on systems with more than a single disk formatted with the older
metadata version 1 format.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>