lib: reset empty zones
If the drive was used with passthrough command tools (e.g. sg3utils or
libzbc) the kernel sequential zone write pointer tracking may be out of
sync with the device. Make sure the kernel state is updated for all
sequential zones by resetting even zones seen as empty as reported by
the device.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>