dev: fix dmz_get_dev_capacity()
The /sys/block/<disk>/queue/nr_zones attribute file was introduced with
kernel 5.0. This causes dmz_get_dev_capacity() to fail to get the number
of zones of the device with older kernels. Fix this by calculating the
number of zones using the capacity and zone size rather than trying to
obtain the same information through sysfs.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>