dmz: format: use unique default label
Using the block device name to generate a dm-zoned device label that
will be used as a DM device name is simple, but has the disadvantage
that if the disk block device name changes, the label may be reused by
some other dm-zoned device and prevent starting the device using the old
block device name as a label.

To avoid this, use the first 8 characters of the DM device UUID to
generate a unique label. In the unlikely event of this label being used
already, fall back to using the device name to create the DM device
name.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>