dmz: Use macro for label and uuid length
Introduce the macros DMZ_LABEL_LEN and DMZ_UUID_LEN to define the
maximum length of labels and uuids and use these macros instead of hard
coded values. This also fixes a bug in the command line interface
preventing labels to be more than 16 characters instead of up allowing
up to 31 characters.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>