dmzadm: Introduce relabel operation
Introduce the --relabel command line operation to allow a user to change
the label of a formatted device. dmz_relabel() is the internal function
implementating this operation. dmz_relabel() checks the device(s) super
blocks first and if no error is detected, the lable of all super blocks
is updated without any other change to the super blocks.

If the --label option is used, the user specified label will be used for
relabelling. Otherwise, the default label is used, as provided by the
new function dmz_get_label(). The default label is the serial number of
the metadata block device prefixed with "dmz-". If the serial number of
the metadata block device is not available, the default label is the
metadata block device name prefixed with "dmz-", as before.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>