dmz: dev: allow overwriting superblocks in sequential zones
When a regular block device is used as cache/metadata device, secondary
zoned block devices can be used even if they do not have any
conventional zones. In this case, the super block is written in the
first sequential zone of the device. This is handled by the format
operation as all zones are reset prior to writing the super block.

However, the relabel and repair operation may overwrite this superblock
to update it. Thus, we need to reset the sequential zone before writing
to avoid write errors. Add the dmz_prepare_super_zone() format helper
function to handle this. Also make sure that secondary devices are
always open with O_DIRECT to avoid doing buffered writes into sequential
zones.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>