dmz: check: Fix checks for multi-device case
When multiple devices are used, the superblock of secondary devices is
always written in the first zone of the device. Since this zone will
never be used to store data, it must be excludeed from mapping and
bitmaps checks.

While at it, also fix dmz_check_sb() to immediately return an error if
reading the superblock fails.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>