configure: use pkg-config
There was a build problem with the update to version 2.2.0 on openSUSE
because on openSUSE the include path of libkmod.h is different than on
i.e. Fedora.

Use pkg-config to obtain the CFLAGS and LIBS for all external libraries so
we're not running into this kind of problem again in the future and/or
don't need a openSUSE specific external patch to fix the build.

Signed-off-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>