configure: use pkg-config
There was a build problem with the update to version 2.2.0 on openSUSE
because on openSUSE the include path of libkmod.h is different than on
i.e. Fedora.

Use pkg-config to obtain the CFLAGS and LIBS compile options for all
external libraries so we are not running into this kind of problem
again in the future. With this change, we do not need a openSUSE
specific external patch to fix the build.

[Damien]
Update the RPM spec file to also dependencies using pkgconfig().

Signed-off-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>